+++ app/models/feed.rb
@@ -1,3 +1,5 @@
 class Feed < ApplicationRecord
+  attr_readonly :space_id
+
   belongs_to :space
 end